// client/src/contexts/socket/socketTypes.ts
// src/constants/socketTypes.js
export enum SOCKET_EVENTS {
    REALTIME_EVENT = 'realtimeEvent', //LẮNG NGHE
    SET_SOCKET = 'SET_SOCKET', //CONNECT
}

export enum SOCKET_PAYLOAD_TYPES {
    MESSAGE = 'MESSAGE',
    TYPING = 'TYPING',
    COMMENT = 'COMMENT',
    GROUP_UPDATE = 'GROUP_UPDATE',
    NOTIFICATION = 'NOTIFICATION',
    STOP_TYPING = 'STOP_TYPING',
    READ_RECEIPT = 'READ_RECEIPT',
    SET_ACTIVE_CONVERSATION = 'SET_ACTIVE_CONVERSATION',
}